+++ GuardiansoftheSpectrum/app/src/main/java/edu/colorado/gots/guardiansofthespectrum/MainActivity.java
@@ -4,14 +4,11 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
-import android.view.View;
-import android.widget.CompoundButton;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.view.MenuItem;
@@ -21,11 +18,8 @@ public class MainActivity extends AppCompatActivity {
     Switch serviceSwitch;
     Intent serviceIntent;
 
-    /*
-    private boolean scanEnabled = true;
-    private CounterReceiver counterReceiver;
+    protected boolean scanEnabled = true;
     private BatteryReceiver batteryReceiver;
-*/
 
 
     /**
@@ -39,18 +33,13 @@ public class MainActivity extends AppCompatActivity {
         Toolbar mToolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(mToolbar);
 
-        /*
-        //set up a quick and dirty listener to receiver counter updates from the service
-        counterReceiver = new CounterReceiver();
-        LocalBroadcastManager.getInstance(this).registerReceiver(counterReceiver,
-                                                                 new IntentFilter(ScanService.GOTS_COUNTER));
         //set up a listener to manage low-battery notifications
         batteryReceiver = new BatteryReceiver();
         IntentFilter intentFilter = new IntentFilter();
         intentFilter.addAction(Intent.ACTION_BATTERY_LOW);
         intentFilter.addAction(Intent.ACTION_BATTERY_OKAY);
         registerReceiver(batteryReceiver, new IntentFilter());
-        */
+
 
         mTextView = (TextView)findViewById(R.id.textview1);
         mTextView.setText(R.string.Hello);
@@ -89,31 +78,11 @@ public class MainActivity extends AppCompatActivity {
 
         }
     }
-}
-
-
-
-/*
     public void onDestroy() {
-        LocalBroadcastManager.getInstance(this).unregisterReceiver(counterReceiver);
         unregisterReceiver(batteryReceiver);
         super.onDestroy();
     }
 
-    //start the scan activity after we get a button click
-    public void initiateScan(View v) {
-        Intent i = new Intent(this, ScanActivity.class);
-        startActivity(i);
-    }
-
-
-
-    public class CounterReceiver extends BroadcastReceiver {
-        public void onReceive(Context context, Intent intent){
-            serviceSwitch.setText(String.format("Service running: %d\n",
-                                                intent.getIntExtra(ScanService.GOTS_COUNTER_EXTRA, 0)));
-        }
-    }
 
 
     private class BatteryReceiver extends BroadcastReceiver  {
@@ -127,9 +96,11 @@ public class MainActivity extends AppCompatActivity {
             }
         }
     }
-
 }
-*/
+
+
+
+
