+++ GuardiansoftheSpectrum/app/src/main/java/edu/colorado/gots/guardiansofthespectrum/ScanService.java
@@ -31,13 +31,11 @@ public class ScanService extends IntentService {
             System.out.println("Running background Scan Service\n");
             try {
                 Thread.sleep(1000);
-                /*
                 counter++;
-                Intent broadcast = new Intent(this, MainActivity.CounterReceiver.class);
+                Intent broadcast = new Intent(this, SettingsActivity.CounterReceiver.class);
                 broadcast.setAction(GOTS_COUNTER);
                 broadcast.putExtra(GOTS_COUNTER_EXTRA, counter);
                 LocalBroadcastManager.getInstance(this).sendBroadcast(broadcast);
-            */
 
             } catch (InterruptedException e) {
                 Thread.currentThread().interrupt();
