+++ GuardiansoftheSpectrum/app/src/main/res/layout/activity_scan.xml
@@ -10,12 +10,19 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="edu.colorado.gots.guardiansofthespectrum.ScanActivity">
 
-    <TextView
-        android:id="@+id/wifi_scanStat"
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:scrollbars="vertical"/>
+        android:layout_height="match_parent"
+        android:layout_centerVertical="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginStart="35dp">
 
+        <TextView
+            android:id="@+id/wifi_scanStat"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:scrollbars="vertical"/>
+    </ScrollView>
 
 
 </RelativeLayout>
